+++ src/main/java/tornadofx/Nodes.kt
@@ -391,48 +391,6 @@ class TableColumnCellCache<T>(private val cacheProvider: (T) -> Node) {
     fun getOrCreateNode(value: T) = store.getOrPut(value, { cacheProvider(value) })
 }
 
-/**
- * Calculate a unique Node per item and set this Node as the graphic of the TableCell.
- *
- * To support this feature, a custom cellFactory is automatically installed, unless an already
- * compatible cellFactory is found. The cellFactories installed via #cellFormat already knows
- * how to retrieve cached values.
- */
-fun <S, T> TableColumn<S, T>.cellCache(cachedGraphicProvider: (T) -> Node) {
-    properties["tornadofx.cellCache"] = TableColumnCellCache(cachedGraphicProvider)
-    // Install a cache capable cellFactory it none is present. The default cellFormat factory will do.
-    if (properties["tornadofx.cellCacheCapable"] != true) {
-        cellFormat { }
-    }
-}
-
-@Suppress("UNCHECKED_CAST")
-fun <S, T> TableColumn<S, T>.cellFormat(formatter: TableCell<S, T>.(T) -> Unit) {
-    properties["tornadofx.cellCacheCapable"] = true
-    cellFactory = Callback { _: TableColumn<S, T> ->
-        object : TableCell<S, T>() {
-            override fun updateItem(item: T, empty: Boolean) {
-                super.updateItem(item, empty)
-
-                if (item == null || empty) {
-                    textProperty().unbind()
-                    graphicProperty().unbind()
-                    text = null
-                    graphic = null
-                } else {
-                    // Consult the cell cache before calling the formatter function
-                    val cellCache = this@cellFormat.properties["tornadofx.cellCache"]
-                    if (cellCache is TableColumnCellCache<*>) {
-                        graphicProperty().unbind()
-                        graphic = (cellCache as TableColumnCellCache<T>).getOrCreateNode(item)
-                    }
-                    formatter(this, item)
-                }
-            }
-        }
-    }
-}
-
 fun <S, T> TableColumn<S, T>.cellDecorator(decorator: TableCell<S, T>.(T) -> Unit) {
     val originalFactory = cellFactory
 
@@ -855,10 +813,6 @@ fun <T> TableView<T>.regainFocusAfterEdit() = apply {
 fun <T, S : Any> TableColumn<T, S>.makeEditable(converter: StringConverter<S>): TableColumn<T, S> = apply {
     tableView?.isEditable = true
     cellFactory = TextFieldTableCell.forTableColumn<T, S>(converter)
-}
-
-fun <T, S : Any> TableColumn<T, S?>.converter(converter: StringConverter<S>): TableColumn<T, S?> = apply {
-    cellFormat { text = converter.toString(it) }
 }
 
 fun <T> TreeTableView<T>.populate(itemFactory: (T) -> TreeItem<T> = { TreeItem(it) }, childFactory: (TreeItem<T>) -> Iterable<T>?) =