+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Additions
 
+- TableColumn.cellFragment to match ListView.cellFragment + SmartTableCell which encapsulate cellFormat, cellCache and cellFragment
 - bindChildren(observableSet, converter) to completement the observableList version
 - sequentialTransition, parallelTransition builders (https://github.com/edvin/tornadofx/issues/373)
 - ObservableList<*>.sizeProperty keeps track of the number of items in an ObservableList
@@ -26,6 +27,7 @@
 
 - Kotlin 1.1.3-2
 - DataGrid receives focus on click
+- TableView refactoring, all cell manipulation functions are encapsulated in a SmartTableCell
 
 ## [1.7.8] - 2017-06-25